refractored oppportunity validators
